chore: configure ESLint and Prettier
