+++ .prettierrc.js
@@ -1,0 +1,6 @@
+export default {
+  trailingComma: 'all',
+  singleQuote: true,
+  printWidth: 120,
+  tabWidth: 2,
+};
