+++ package.json
@@ -2,9 +2,9 @@
   "name": "taqtile-api",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "start": "node src/server.js",
-    "dev": "tsx --watch src/index.ts",
+    "dev": "tsx --watch src/server.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -22,7 +22,14 @@
     "type-graphql": "2.0.0-rc.2"
   },
   "devDependencies": {
+    "@eslint/js": "9.11.1",
+    "@types/eslint__js": "8.42.3",
+    "eslint": "9.11.1",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-prettier": "5.2.1",
+    "prettier": "3.3.3",
     "tsx": "4.19.1",
-    "typescript": "5.6.2"
+    "typescript": "5.5.3",
+    "typescript-eslint": "8.8.0"
   }
 }
