+++ src/resolvers/hello-world-resolver.ts
@@ -1,0 +1,9 @@
+import { Query, Resolver } from 'type-graphql';
+
+@Resolver()
+export class HelloWorld {
+  @Query(() => String)
+  async HelloWorld() {
+    return 'Hello, World!';
+  }
+}
