--- src/resolvers/helloWorld.resolver.ts
@@ -1,11 +1,0 @@
-import { Query, Resolver } from "type-graphql";
-
-@Resolver()
-export class HelloWorld {
-
-    @Query(() => String)
-    async HelloWorld(){
-        return "Hello, World!"
-    }
-
-}
